Move to spectator

+++ priam/src/main/java/com/netflix/priam/google/GoogleEncryptedFileSystem.java
@@ -315,7 +315,7 @@ public class GoogleEncryptedFileSystem implements IBackupFileSystem, GoogleEncry
     }
 
     @Override
-    public int getAWSSlowDownExceptionCounter() {
+    public long getAWSSlowDownExceptionCounter() {
         return 0;
     }
